# webroot/swagger/js/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Wista-S3",
    "description": "将 Wistia 上的视频上传至 S3。",
    "version": "V1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": [],
  "paths": {
    "/refresh/media": {
      "post": {
        "tags": [],
        "summary": "从S3刷新所有 uploaded video 到 local DB",
        "description": "<p>从S3刷新所有 uploaded video 到 local DB。同时拉取 index.json（视频基本信息）和 index-ai.json（AI 索引结果）。</p>",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          }
        }
      }
    },
    "/sync/wistia": {
      "post": {
        "tags": [],
        "summary": "從 Wistia API 同步所有視頻資訊到本地資料庫",
        "description": "<p>從 Wistia API 拉取所有 active + archived 視頻的 metadata，快取到本地 BoltDB。使用 async task 模式，需輪詢 GET /tasks/{id} 取得結果。</p>",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          }
        }
      }
    },
    "/wistia/media": {
      "get": {
        "tags": [],
        "summary": "查詢 Wistia 視頻目錄（從本地快取）",
        "description": "<p>從本地 BoltDB 快取中查詢 Wistia 視頻目錄，支援依 hash 查詢單一視頻、依 archived 狀態篩選，以及分頁。</p>",
        "parameters": [
          {
            "name": "hash",
            "in": "query",
            "description": "視頻 HashId（提供時回傳單一視頻）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "archived",
            "in": "query",
            "description": "依 archived 狀態篩選（true 或 false）",
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "頁碼（預設 1）",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "每頁筆數（預設 50，最大 100）",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/WistiaRespVideo"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": { "type": "integer" },
                        "per_page": { "type": "integer" },
                        "total": { "type": "integer" },
                        "total_pages": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "未找到（hash 查詢無結果）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          },
          "500": {
            "description": "錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          }
        }
      }
    },
    "/media": {
      "get": {
        "tags": [],
        "summary": "获取所有 uploaded 视频信息",
        "description": "<p>获取所有 uploaded 视频信息</p>",
        "parameters": [
          {
            "name": "hash",
            "in": "query",
            "description": "视频 HashId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          { "$ref": "#/components/schemas/WistiaRespVideo" },
                          {
                            "type": "object",
                            "properties": {
                              "index": {
                                "$ref": "#/components/schemas/DashScopeIndexResult",
                                "description": "AI 索引結果（若存在）"
                              }
                            }
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "tags": [],
        "summary": "获取任务状态",
        "description": "<p>获取任务状态</p>",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "任务ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          },
          "500": {
            "description": "錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          }
        }
      }
    },
    "/move/{hash}": {
      "post": {
        "tags": [],
        "summary": "迁移Video到S3",
        "description": "<p>迁移Video到S3</p>",
        "operationId": "single-move",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "Wistia Video HashId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "forceRefresh",
            "in": "query",
            "description": "强制重新迁移，覆盖原来数据",
            "schema": {
              "type": "string",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          }
        }
      }
    },
    "/move": {
      "post": {
        "tags": [],
        "summary": "批量迁移Video到S3",
        "description": "<p>批量迁移Video到S3</p>",
        "operationId": "multiple-move",
        "parameters": [
          {
            "name": "forceRefresh",
            "in": "query",
            "description": "强制重新迁移，覆盖原来数据",
            "schema": {
              "type": "string",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MultipleMediaBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "請求格式錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          },
          "500": {
            "description": "錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          }
        }
      }
    },
    "/index/{hash}": {
      "post": {
        "tags": [],
        "summary": "對單一視頻進行 AI 索引（字幕轉錄 + 摘要 + 章節）",
        "description": "對單一視頻進行 AI 索引。若本地資料庫已有索引結果且 force 不為 true，直接同步回傳 DashScopeIndexResult；否則建立非同步任務回傳 Task，需輪詢 GET /tasks/{id} 取得結果。",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "視頻 HashId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "force",
            "in": "query",
            "description": "強制重新索引，忽略快取",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean"
                    },
                    "data": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/Task"
                        },
                        {
                          "$ref": "#/components/schemas/DashScopeIndexResult"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "請求格式錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          },
          "500": {
            "description": "錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [],
        "summary": "獲取視頻 AI 索引結果",
        "description": "從本地資料庫獲取視頻的 AI 索引結果",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "視頻 HashId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/DashScopeIndexResult"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          },
          "500": {
            "description": "錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          }
        }
      }
    },
    "/index": {
      "post": {
        "tags": [],
        "summary": "批量對視頻進行 AI 索引",
        "description": "批量提交視頻 AI 索引任務",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MultipleMediaBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "請求格式錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          },
          "500": {
            "description": "錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStandardError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "APIResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean"
          },
          "data": {}
        }
      },
      "APIStandardError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "init",
              "running",
              "finished",
              "error"
            ]
          },
          "result": {}
        }
      },
      "MoveToS3Result": {
        "type": "object",
        "properties": {
          "hash": {
            "type": "string"
          },
          "cloudfront": {
            "type": "string"
          },
          "s3": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "MultipleMediaBody": {
        "type": "object",
        "properties": {
          "media": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "視頻HashId"
            }
          }
        }
      },
      "WistiaRespVideo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "hashed_id": {
            "type": "string"
          },
          "duration": {
            "type": "number",
            "format": "float"
          },
          "status": {
            "type": "string"
          },
          "progress": {
            "type": "number",
            "format": "float"
          },
          "archived": {
            "type": "boolean"
          },
          "section": {
            "type": "string"
          },
          "thumbnail": {
            "$ref": "#/components/schemas/WistiaRespVideoThumbnail"
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WistiaRespVideoAsset"
            }
          },
          "project": {
            "$ref": "#/components/schemas/WistiaRespVideoProject"
          }
        }
      },
      "WistiaRespVideoAsset": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "fileSize": {
            "type": "integer"
          },
          "contentType": {
            "type": "string"
          },
          "width": {
            "type": "integer"
          },
          "height": {
            "type": "integer"
          }
        }
      },
      "WistiaRespVideoThumbnail": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "width": {
            "type": "integer"
          },
          "height": {
            "type": "integer"
          }
        }
      },
      "WistiaRespVideoProject": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "hashed_id": {
            "type": "string"
          }
        }
      },
      "DashScopeIndexResult": {
        "type": "object",
        "properties": {
          "hashId": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/WistiaRespVideoAsset"
          },
          "generatedAt": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "subtitles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashScopeSubtitleEntry"
            }
          },
          "chapters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashScopeChapterEntry"
            }
          },
          "tokenUsage": {
            "$ref": "#/components/schemas/DashScopeTokenUsage"
          }
        }
      },
      "DashScopeSubtitleEntry": {
        "type": "object",
        "properties": {
          "start": {
            "type": "number",
            "format": "double"
          },
          "end": {
            "type": "number",
            "format": "double"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "DashScopeChapterEntry": {
        "type": "object",
        "properties": {
          "start": {
            "type": "number",
            "format": "double"
          },
          "end": {
            "type": "number",
            "format": "double"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "DashScopeTokenUsage": {
        "type": "object",
        "properties": {
          "inputK": {
            "type": "number"
          },
          "outputK": {
            "type": "number"
          },
          "totalK": {
            "type": "number"
          }
        }
      }
    }
  }
}